requirements.txt
# requirements.txt
# -----------------
# Все основные зависимости бота c зафиксированными версиями,
# протестированными в ветке develop (июнь 2025).


python-telegram-bot==22.1     # Telegram Bot API wrapper (async / PTB v20)
openai==1.23.0                # Новое официальное OpenAI-SDK (AsyncOpenAI)
httpx==0.27.0                 # Асинхронные HTTP-запросы (используется внутри PTB / openai)
python-dotenv==1.0.1          # Подхватываем переменные из .env
aiofiles==23.2.1              # Асинхронное чтение картинок перед отправкой

# --- опционально, но полезно ---
# python-slugify==8.0.4       # если будете генерировать «чистые» названия файлов
# pytest==8.1.1               # юнит-тесты (планы на GitHub CI)

# ⚠️ После добавления файла в репозиторий не забудьте выполнить:
# pip install -r requirements.txt
# чтобы локальное окружение соответствовало продакшен-зависимостям.